+++ tests/unit/test_config.py
@@ -1,18 +1,19 @@
 import os
+import sys
 import unittest
 
 from kuyruk.config import Config
-from kuyruk.master import parse_queues_str
-from kuyruk.test.unit import config as user_config
+import config as user_config
+
+sys.path.append(os.path.abspath(os.path.dirname(__file__)))
 
 
 class ConfigTestCase(unittest.TestCase):
 
     def test_from_pymodule(self):
-        dirname = os.path.dirname(__file__)
         config = Config()
-        config.from_pymodule('kuyruk.test.unit.config')
-        self.assertEqual(config.REDIS_DB, 2)
+        config.from_pymodule('config2')
+        self.assertEqual(config.MAX_LOAD, 20)
 
     def test_from_pyfile(self):
         dirname = os.path.dirname(__file__)
@@ -26,7 +27,3 @@ class ConfigTestCase(unittest.TestCase):
         config = Config()
         config.from_object(user_config)
         self.assertEqual(config.MAX_LOAD, 21)
-
-    def test_queues_string(self):
-        assert parse_queues_str(" a,b, c, 2*d, e*3, 2*@f ") == \
-            ['a', 'b', 'c', 'd', 'd', 'e', 'e', 'e', '@f', '@f']